requirements.txt
aiomysql==0.1.1
aiosmtplib==1.1.7
aiosqlite==0.17.0
anyio==3.6.1
bcrypt==4.0.0
blinker==1.5
click==8.1.3
databases==0.6.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.82.0
fastapi-mail==1.1.4
greenlet==1.1.3
h11==0.13.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.10.2
PyMySQL==1.0.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
